+++ core/pxstring.cpp
@@ -43,7 +43,7 @@ void string::stralloc(const char * sTxt, const size_t szLen)
 	m_sData = (char *)malloc(m_szDataLen + 1);
 	if (m_sData == NULL)
 	{
-		AppendDebugLogFormat("[MEM] Cannot allocate %" PRIu64 " bytes for sData in string::stralloc\n", (uint64_t)(m_szDataLen + 1));
+		AppendDebugLogFormat("[MEM] Cannot allocate %zu bytes for sData in string::stralloc\n", m_szDataLen+1);
 		
 		return;
 	}
@@ -87,7 +87,7 @@ string::string(const uint32_t ui32Number) : m_sData((char *)sEmpty), m_szDataLen
 	int iLen = snprintf(tmp, 16, "%u", ui32Number);
 	if (iLen > 0)
 	{
-		stralloc(tmp, iLen);
+	stralloc(tmp, iLen);
 	}
 #endif
 }
@@ -103,7 +103,7 @@ string::string(const int32_t i32Number) : m_sData((char *)sEmpty), m_szDataLen(0
 	int iLen = snprintf(tmp, 16, "%d", i32Number);
 	if (iLen > 0)
 	{
-		stralloc(tmp, iLen);
+	stralloc(tmp, iLen);
 	}
 #endif
 }
@@ -119,7 +119,7 @@ string::string(const uint64_t ui64Number) : m_sData((char *)sEmpty), m_szDataLen
 	int iLen = snprintf(tmp, 32, "%" PRIu64, ui64Number);
 	if (iLen > 0)
 	{
-		stralloc(tmp, iLen);
+	stralloc(tmp, iLen);
 	}
 #endif
 }
@@ -135,7 +135,7 @@ string::string(const int64_t i64Number) : m_sData((char *)sEmpty), m_szDataLen(0
 	int iLen = snprintf(tmp, 32, "%" PRId64, i64Number);
 	if (iLen > 0)
 	{
-		stralloc(tmp, iLen);
+	stralloc(tmp, iLen);
 	}
 #endif
 }
@@ -154,7 +154,7 @@ string::string(const string & sStr1, const string & sStr2) : m_sData((char *)sEm
 	m_sData = (char *)malloc(m_szDataLen + 1);
 	if (m_sData == NULL)
 	{
-		AppendDebugLogFormat("[MEM] Cannot allocate %" PRIu64 " bytes for sData in string::string(string, string)\n", (uint64_t)(m_szDataLen + 1));
+		AppendDebugLogFormat("[MEM] Cannot allocate %zu bytes for sData in string::string(string, string)\n", m_szDataLen+1);
 		
 		return;
 	}
@@ -179,7 +179,7 @@ string::string(const char * sTxt, const string & sStr) : m_sData((char *)sEmpty)
 	m_sData = (char *)malloc(m_szDataLen + 1);
 	if (m_sData == NULL)
 	{
-		AppendDebugLogFormat("[MEM] Cannot allocate %" PRIu64 " bytes for sData in string::string(char, string)\n", (uint64_t)(m_szDataLen + 1));
+		AppendDebugLogFormat("[MEM] Cannot allocate %zu bytes for sData in string::string(char, string)\n", m_szDataLen+1);
 		
 		return;
 	}
@@ -204,7 +204,7 @@ string::string(const string & sStr, const char * sTxt) : m_sData((char *)sEmpty)
 	m_sData = (char *)malloc(m_szDataLen + 1);
 	if (m_sData == NULL)
 	{
-		AppendDebugLogFormat("[MEM] Cannot allocate %" PRIu64 " bytes for sData in string::string(string, char)\n", (uint64_t)(m_szDataLen + 1));
+		AppendDebugLogFormat("[MEM] Cannot allocate %zu bytes for sData in string::string(string, char)\n", m_szDataLen+1);
 		
 		return;
 	}
@@ -288,7 +288,7 @@ string & string::operator+=(const char * sTxt)
 	{
 		m_sData = oldbuf;
 		
-		AppendDebugLogFormat("[MEM] Cannot allocate %" PRIu64 " bytes for sData in string::operator+=(char)\n", (uint64_t)(m_szDataLen + szLen + 1));
+		AppendDebugLogFormat("[MEM] Cannot allocate %zu bytes for sData in string::operator+=(char)\n", m_szDataLen+szLen+1);
 		
 		return *this;
 	}
@@ -323,7 +323,7 @@ string & string::operator+=(const string & sStr)
 	{
 		m_sData = oldbuf;
 		
-		AppendDebugLogFormat("[MEM] Cannot allocate %" PRIu64 " bytes for sData in string::operator+=(string)\n", (uint64_t)(m_szDataLen + sStr.size() + 1));
+		AppendDebugLogFormat("[MEM] Cannot allocate %zu bytes for sData in string::operator+=(string)\n", m_szDataLen+sStr.size()+1);
 		
 		return *this;
 	}
